requirements.txt
flask
python-dotenv
weaviate-client
openai
aiohttp
anthropic
flask[async]
anthropic